feat: Formularios de Login y Registro Se agregaron formularios de Login y Registro al sistema. Se restringieron funcionalidades ABM para que solo puedan realizarse si el usuario inicio sesion.

+++ public/delete_user.php
@@ -9,6 +9,15 @@ $userManager = new UserManager($dataBaseConnection);
 
 $id = $_GET['id'];
 
+if (session_status() == PHP_SESSION_NONE) {
+    session_start();
+}
+
+if (!isset($_SESSION['user_id'])) {
+    header("Location: login.php");
+    exit();
+}
+
 if ($userManager -> deleteUser($id)) {
     echo "<div>Usuario eliminado correctamente</div>";
 } else {
